Support bundling all transitive crates in one directory
Transitive crate dependencies are passed to rustc by specifying the
folder they live in as a library search path (an `-Ldependency=path`
flag). Given that bazel builds all libraries into their own folders,
this means passing potentially hundreds of paths to rustc.

For a number of transitive dependencies greater than 400 we've observed
that rustc runs into some odd `LoadLibraryExW` error and fails to find
some transitive crates and thus fails to build.

For this reason, this patch can optionally createc symlinks to all the
transitive crates into a single directory that it then sets as the
`-Ldependency=` flag to rustc. We only do this when building on Windows
and when the number of transitive crates is greater than 400, which
happens for a handful of top-level crates in the dropbox_core project.